+++ bin/params.ml
@@ -43,9 +43,9 @@ let hash_chunk_size = 4096
 
 let text_editor = ref ""
 
-let default_sample_count_per_document = 5
+let default_samples_per_document = 5
 
-let sample_count_per_document = ref default_sample_count_per_document
+let samples_per_document = ref default_samples_per_document
 
 type style_mode = [ `Never | `Always | `Auto ]